Move junit finder classes to serenity_bdd package
