+++ serenity-junit/src/main/java/net/serenity_bdd/junit/finder/FinderType.java
@@ -1,4 +1,4 @@
-package net.thucydides.junit.finder;
+package net.serenity_bdd.junit.finder;
 
 public enum FinderType {
     ALL_TESTS,
