+++ serenity-junit/src/main/java/net/serenity_bdd/junit/finder/TestFinderBuilderFactory.java
@@ -1,4 +1,4 @@
-package net.thucydides.junit.finder;
+package net.serenity_bdd.junit.finder;
 
 
 public class TestFinderBuilderFactory {
