+++ serenity-junit/src/main/java/net/thucydides/junit/listeners/TestCountListener.java
@@ -11,7 +11,7 @@ import net.thucydides.core.steps.ExecutedStepDescription;
 import net.thucydides.core.steps.StepFailure;
 import net.thucydides.core.steps.StepListener;
 import net.thucydides.core.util.EnvironmentVariables;
-import net.thucydides.junit.finder.TestFinder;
+import net.serenity_bdd.junit.finder.TestFinder;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
